add cookbook

+++ springboot/src/main/java/com/cztek/springboot/controller/DeleteCookListController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import com.cztek.springboot.Util.DateFormat;
 import com.cztek.springboot.service.IUserBookService;
 
 @Controller
@@ -23,8 +24,14 @@ public class DeleteCookListController {
 	@ResponseBody
 	public Map<String,Integer> deleteUserBookId(@RequestParam(value="data") Integer userBookId) {
 		Map<String, Integer> map = new HashMap<>();
-	    int deleteUserBook = userBookService.deleteUserBook(userBookId);
-		map.put("message",deleteUserBook);
+		String time = DateFormat.currentTime("17:00:00");
+		boolean resultInt = DateFormat.resultInt(time);
+		if (resultInt == true) {
+		int deleteUserBook = userBookService.deleteUserBook(userBookId);
+		map.put("message", deleteUserBook);
+		}else{
+			map.put("message",-1);
+		}
 		return map;
 	}
 }
